requirements.txt
# Core Transcription & Audio
faster-whisper
pyaudio
numpy
moviepy
jumpcutter
numpy<1.24

# Note Generation & Translation
google-generative-ai
markdown

# Utilities
python-dotenv